+++ Assets/02_Script/Environment/Growable.cs
@@ -39,29 +39,10 @@ public class Growable : TimeAgent
         UpdateSprite();
     }
 
-    private void OnEnable()
+    protected override void OnEnable()
     {
-        if (EnvironmentManager.Instance.TryGetComponent(out TimeController timeController))
-        {
-            timeController.Subscribe(this);
-        }
+        base.OnEnable();
         ResetTimer();
-    }
-
-    private void OnDisable()
-    {
-        if (isQuitting) { return; }
-        if (EnvironmentManager.Instance.TryGetComponent(out TimeController timeController))
-        {
-            timeController.Unsubscribe(this);
-        }
-    }
-
-    // 게임뷰에서 씬뷰로 넘어올 때 에러 방지
-    bool isQuitting = false;
-    private void OnApplicationQuit()
-    {
-        isQuitting = true;
     }
 
     public override void UpdateTimer()
