+++ Assets/02_Script/Item/Torch.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 588cef802f47e824d9abb2478dba1c9e
