+++ Assets/05_ScriptableObjects/NatureResources/Minerals/DesertMineral2.asset
@@ -19,6 +19,6 @@ MonoBehaviour:
   height: 2
   maxHealth: 100
   dropItems:
-  - data: {fileID: 11400000, guid: e0374df419f3fd74fa0e8d09ee3aa0a1, type: 2}
+  - data: {fileID: 11400000, guid: 32b8e6989fe545f4b8d99985fb258e90, type: 2}
     minAmount: 3
     maxAmount: 5
