+++ Assets/05_ScriptableObjects/NatureResources/Minerals/DesertMineral3.asset
@@ -19,6 +19,6 @@ MonoBehaviour:
   height: 2
   maxHealth: 100
   dropItems:
-  - data: {fileID: 11400000, guid: 1683c1cf70666bb43997a2277d4ec0e3, type: 2}
+  - data: {fileID: 11400000, guid: 0b3e4601311f0294c830cfeefea8d93a, type: 2}
     minAmount: 3
     maxAmount: 5
