+++ Assets/05_ScriptableObjects/NatureResources/Trees/GrassTree1.asset
@@ -19,7 +19,7 @@ MonoBehaviour:
   height: 2
   maxHealth: 100
   dropItems:
-  - data: {fileID: 11400000, guid: 4f1dfe3c1158a8748a54e2cf2413bbdb, type: 2}
+  - data: {fileID: 11400000, guid: 62b608b058674f348a329c683464ff7a, type: 2}
     minAmount: 3
     maxAmount: 5
   timeToAllGrown: 0
