+++ Assets/05_ScriptableObjects/NatureResources/Trees/GrassTree1Trunk.asset
@@ -19,6 +19,6 @@ MonoBehaviour:
   height: 2
   maxHealth: 50
   dropItems:
-  - data: {fileID: 11400000, guid: 4f1dfe3c1158a8748a54e2cf2413bbdb, type: 2}
+  - data: {fileID: 11400000, guid: 62b608b058674f348a329c683464ff7a, type: 2}
     minAmount: 1
     maxAmount: 2
